+++ app/api/teams/route.js
@@ -1,5 +1,7 @@
 import { NextResponse } from "next/server"
 import prisma from "../../libs/prismadb";
+import getCurrentUser from "../../get/server/getCurrentUser";
+
 
 export async function GET(request) {
   try {
@@ -26,9 +28,30 @@ export async function GET(request) {
       },
       include: {
         pool: true,
-        matches1: true,
-        matches2: true,
-        workMatches: true,
+        matches1: {
+          include: {
+            team1: true,
+            team2: true,
+            workTeam: true,
+            sets: true
+          }
+        },
+        matches2: {
+          include: {
+            team1: true,
+            team2: true,
+            workTeam: true,
+            sets: true
+          }
+        },
+        workMatches: {
+          include: {
+            team1: true,
+            team2: true,
+            workTeam: true,
+            sets: true
+          }
+        },
       }
     })
 
@@ -43,12 +66,24 @@ export async function GET(request) {
   }
 }
 
-export async function POST(request, { params }) {
+export async function POST(request) {
+
+  let poolID = n = "";
+  try {
+    const body = await request.json()
+    n = body.name;
+    poolID = body.poolID;
+  } catch (e) {
+    return new NextResponse('Need Body', { status: 400 })
+  }
 
   try {
-    const { poolID, name } = await request.json();
+    const currentUser = await getCurrentUser();
 
-    if (!name) {
+    if (!currentUser?.id || !currentUser?.email || !currentUser?.role === 'admin') {
+      return new NextResponse("Unauthorizied", { status: 401 })
+    }
+    if (!n) {
       return new NextResponse("Need Team Name", { status: 400 })
     }
     if (!poolID) {
@@ -66,7 +101,7 @@ export async function POST(request, { params }) {
     const index = (pool) ? pool.teams.length : 0;
     const team = await prisma.team.create({
       data: {
-        name: name,
+        name: n,
         index: index,
         pool: {
           connect: {
@@ -86,6 +121,12 @@ export async function POST(request, { params }) {
 export async function DELETE(request) {
   try {
     const id = request.nextUrl.searchParams.get("id");
+
+    const currentUser = await getCurrentUser();
+
+    if (!currentUser?.id || !currentUser?.email || !currentUser?.role === 'admin') {
+      return new NextResponse("Unauthorizied", { status: 401 })
+    }
 
     if (!id) {
       const teams = await prisma.team.deleteMany({})
